Add compiled files
Nix and node2nix make it hard to include a build-step in node dependencies that
are not hosted on npmjs. I eventually figured out that I could add pre- and
postInstall hooks to npm2nix's buildNodeDependencies

```diff
3 files changed, 9 insertions(+)
assets/default.nix       | 1 +
assets/node-env.nix      | 4 ++++
assets/node-packages.nix | 4 ++++

modified   assets/default.nix
@@ -14,4 +14,5 @@ in
 import ./node-packages.nix {
   inherit (pkgs) fetchurl nix-gitignore stdenv lib fetchgit;
   inherit nodeEnv;
+  globalBuildInputs = [pkgs.yarn];
 }
modified   assets/node-env.nix
@@ -604,6 +604,8 @@ let
         passAsFile = [ "includeScript" "pinpointDependenciesScript" ];

         installPhase = ''
+          runHook preInstall
+
           source ${installPackage}

           mkdir -p $out/${packageName}
@@ -634,6 +636,8 @@ let

           mv ${packageName} lib
           ln -s $out/lib/node_modules/.bin $out/bin
+
+          runHook postInstall
         '';
       } // extraArgs);

modified   assets/node-packages.nix
@@ -1573,6 +1573,10 @@ let
     production = false;
     bypassCache = true;
     reconstructLock = false;
+    postInstall = ''
+      cd lib/node_modules/@editorjs/table && yarn build
+    '';
+
   };
 in
 {
```

However, the next problem is that the table component uses vite to build the
project with, and vite is not packaged for Nix. After much back and forth, I
decided that the most straightforward way to proceed is just to include the
generated files in the repo.